trading_bot/requirements.txt
Flask==3.0.0
click==8.1.7
python-binance==1.0.19
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
